moved app listen to "server" file to fix jest tests not ending
